+++ eth_defi/vault/risk.py
@@ -271,6 +271,7 @@ _BROKEN_VAULT_CONTRACTS = {
     "0xbF29685856FAE1e228878DFB35B280C0adCC3B05",  # Age old mainnet contract
     "0x802C31C7c0A89D3d289Ca91cD57892d8338BC9cE",  # Age old mainnet contract
     "0x3D75bA301AeB7b707796b3948c1ef20d673a3181",  # Age old mainnet contract
+    "0x5c9674C1c5dB468C9f8f89aF67AE58be1746DE05",  # Age old mainnet contract
     "0x6a6E4ad4a5ca14B940Cd6949b1A90f947AE21c19",  # Broken Gains vault on Berachain - its open PnL feed contract (0x5705554B) causes multicall failures
     "0x5705554BAa86Da01fF4A82d29a1598c5B3A8B476",  # Open PnL feed helper contract for broken Gains vault on Berachain
     "0x8fF6aDBC653405245B6b686E31b14A7da7000281",  # BNB broken contract